rename classes to fit internal name
