+++ x-pack/plugin/core/src/test/java/org/elasticsearch/xpack/core/ml/action/GetMlAutoscalingStatsResponseTests.java
@@ -9,13 +9,13 @@ package org.elasticsearch.xpack.core.ml.action;
 
 import org.elasticsearch.common.io.stream.Writeable;
 import org.elasticsearch.test.AbstractWireSerializingTestCase;
-import org.elasticsearch.xpack.core.ml.action.GetMlAutoscalingResources.Response;
+import org.elasticsearch.xpack.core.ml.action.GetMlAutoscalingStats.Response;
 
 import java.io.IOException;
 
 import static org.elasticsearch.xpack.core.ml.autoscaling.AutoscalingResourcesTests.randomAutoscalingResources;
 
-public class GetMlAutoscalingResourcesResponseTests extends AbstractWireSerializingTestCase<Response> {
+public class GetMlAutoscalingStatsResponseTests extends AbstractWireSerializingTestCase<Response> {
 
     @Override
     protected Writeable.Reader<Response> instanceReader() {
